+++ Locale/de_DE/translations.php
@@ -7,7 +7,7 @@ return array(
     'Adds features to do stuff with subtasks' => 'Fügt Features hinzu, um Dinge mit Teilaufgaben zu tun',
     'Subtask from task times' => 'Teilaufgabe von Aufgaben-Zeiten',
     'Subtask name' => 'Teilaufgabe Name',
-    'Enter name for new subtask, which will get the task times' => 'Name eingenen für die neue Teilaufgabe, die die Zeiten von der Aufgabe übernehmen wird',
+    'Enter name for the combining subtask, which will combine done or tarted work of other subtasks. Done subtasks will be sumarized in this and be deleted afterwards. Started yet not done subtasks will add their spent time to the combined subtask and leave their remaining time.' => 'Name für die kombinierende Teilaufgabe eingeben, die alle erledigten Teilaufgaben-Zeiten summiert bekommt. Erledigte Subtasks werden darin summiert und danach selbst gelöscht. Gestartete und noch nicht erledigte Subtasks fügen ihre getätigte Zeit im neuen kombinierten Subtask hinzu und behalten ihre Restzeit.',
     'Unable to convert to subtask' => 'Kann nicht zur Teilaufgabe konvertieren',
     'Subtask converted from task' => 'Teilaufgabe von Aufgabe konvertiert',
     'Task already has subtasks' => 'Aufgabe hat bereits Teilaufgaben',